REsolve dependency cycle in AMD
