+++ src/vs/workbench/contrib/positronConnections/browser/components/newConnectionModalDialog.tsx
@@ -13,6 +13,7 @@ import 'vs/css!./newConnectionModelDialog';
 import { ContentArea } from 'vs/workbench/browser/positronComponents/positronModalDialog/components/contentArea';
 import { CreateConnection } from 'vs/workbench/contrib/positronConnections/browser/components/newConnectionModalDialog/createConnectionState';
 import { ListDrivers } from 'vs/workbench/contrib/positronConnections/browser/components/newConnectionModalDialog/listDriversState';
+import { Driver } from 'vs/workbench/contrib/positronConnections/browser/components/newConnectionModalDialog/driver';
 
 const NEW_CONNECTION_MODAL_DIALOG_WIDTH = 700;
 const NEW_CONNECTION_MODAL_DIALOG_HEIGHT = 430;
@@ -76,44 +77,4 @@ const NewConnectionModalDialog = (props: PropsWithChildren<NewConnectionModalDia
 		</div>
 	</PositronModalDialog>;
 };
-
-export enum InputType {
-	String = 'string',
-	Number = 'number',
-	Boolean = 'boolean',
-}
-
-export interface Input {
-	// The unique identifier for the input.
-	id: string;
-	// A human-readable label for the input.
-	label: string;
-	// The type of the input.
-	type: InputType;
-}
-
-export interface Driver {
-	// The unique identifier for the driver.
-	driverId: string;
-	// The language identifier for the driver.
-	// Drivers are grouped by language, not by runtime.
-	languageId: string;
-	// A human-readable name for the driver.
-	name: string;
-	// The base64-encoded SVG icon for the driver.
-	base64EncodedIconSvg?: string;
-	// The inputs required to create a connection.
-	// For instance, a connection might require a username
-	// and password.
-	inputs: Array<Input>;
-	// Generates the connection code based on the inputs.
-	generateCode?: (inputs: Array<Input>) => string;
-	// Checks if the dependencies for the driver are installed
-	// and functioning.
-	checkDependencies?: () => Promise<boolean>;
-	// Installs the dependencies for the driver.
-	// For instance, R packages would install the required
-	// R packages, and or other dependencies.
-	installDependencies?: () => Promise<boolean>;
-}
